+++ src/tagpack/actorpack_schema.py
@@ -1,4 +1,5 @@
 """ActorPack - A wrappers ActorPack Schema"""
+
 import importlib.resources as pkg_resources
 
 import pandas as pd